Data.Functor.Contravariant.Coyoneda

Description

The co-Yoneda lemma for presheafs states that f is naturally isomorphic to Coyoneda f.

Documentation

data Coyoneda f a where

A Contravariant functor (aka presheaf) suitable for Yoneda reduction.

http://ncatlab.org/nlab/show/Yoneda+reduction

Constructors

Coyoneda :: (a -> b) -> f b -> Coyoneda f a

liftCoyoneda :: f a -> Coyoneda f a

Coyoneda "expansion" of a presheaf

liftCoyoneda . lowerCoyoneda ≡ id
lowerCoyoneda . liftCoyoneda ≡ id

lowerCoyoneda :: Contravariant f => Coyoneda f a -> f a

Coyoneda reduction on a presheaf

hoistCoyoneda :: (forall a. f a -> g a) -> Coyoneda f b -> Coyoneda g b

Lift a natural transformation from f to g to a natural transformation from Coyoneda f to Coyoneda g.