Data.Functor.Coyoneda

Description

The co-Yoneda lemma for a covariant Functor f states that Coyoneda f is naturally isomorphic to f.

Documentation

data Coyoneda f a where

A covariant form suitable for Yoneda reduction

Constructors

Coyoneda :: (b -> a) -> f b -> Coyoneda f a

liftCoyoneda :: f a -> Coyoneda f a

Yoneda expansion

 liftCoyoneda . lowerCoyoneda ≡ id
 lowerCoyoneda . liftCoyoneda ≡ id

 lowerCoyoneda (liftCoyoneda fa) = -- by definition
 lowerCoyoneda (Coyoneda id fa)  = -- by definition
 fmap id fa                      = -- functor law
 fa

 lift = liftCoyoneda

lowerCoyoneda :: Functor f => Coyoneda f a -> f a

Yoneda reduction lets us walk under the existential and apply fmap.

http://ncatlab.org/nlab/show/Yoneda+reduction

You can view Coyoneda as just the arguments to fmap tupled up.

 lower = lowerM = lowerCoyoneda

lowerM :: Monad f => Coyoneda f a -> f a

Yoneda reduction given a Monad lets us walk under the existential and apply liftM.

You can view Coyoneda as just the arguments to liftM tupled up.

 lower = lowerM = lowerCoyoneda

as a Left Kan extension

coyonedaToLan :: Coyoneda f a -> Lan Identity f a

Coyoneda f is the left Kan extension of f along the Identity functor.

 coyonedaToLan . lanToCoyoneda ≡ id
 lanToCoyoneda . coyonedaToLan ≡ id

lanToCoyoneda :: Lan Identity f a -> Coyoneda f a

as a Left Kan lift

coyonedaToLift :: Coyoneda f a -> Lift Identity f a

Coyoneda f is the left Kan lift of f along the Identity functor.

 coyonedaToLift . liftToCoyoneda ≡ id
 liftToCoyoneda . coyonedaToLift ≡ id

liftToCoyoneda :: Functor f => Lift Identity f a -> Coyoneda f a