Data.Functor.Yoneda

Description

The covariant form of the Yoneda lemma states that f is naturally isomorphic to Yoneda f.

This is described in a rather intuitive fashion by Dan Piponi in

http://blog.sigfpe.com/2006/11/yoneda-lemma.html

Documentation

newtype Yoneda f a

Yoneda f a can be viewed as the partial application of fmap to its second argument.

Constructors

Yoneda

Fields

runYoneda :: forall b. (a -> b) -> f b

liftYoneda :: Functor f => f a -> Yoneda f a

The natural isomorphism between f and Yoneda f given by the Yoneda lemma is witnessed by liftYoneda and lowerYoneda

 liftYoneda . lowerYoneda ≡ id
 lowerYoneda . liftYoneda ≡ id

 lowerYoneda (liftYoneda fa) =         -- definition
 lowerYoneda (Yoneda (f -> fmap f a)) -- definition
 (f -> fmap f fa) id                  -- beta reduction
 fmap id fa                            -- functor law
 fa

 lift = liftYoneda

lowerYoneda :: Yoneda f a -> f a

maxF :: (Functor f, Ord (f a)) => Yoneda f a -> Yoneda f a -> Yoneda f a

minF :: (Functor f, Ord (f a)) => Yoneda f a -> Yoneda f a -> Yoneda f a

maxM :: (Monad m, Ord (m a)) => Yoneda m a -> Yoneda m a -> Yoneda m a

minM :: (Monad m, Ord (m a)) => Yoneda m a -> Yoneda m a -> Yoneda m a

as a right Kan extension

yonedaToRan :: Yoneda f a -> Ran Identity f a

Yoneda f can be viewed as the right Kan extension of f along the Identity functor.

 yonedaToRan . ranToYoneda ≡ id
 ranToYoneda . yonedaToRan ≡ id

ranToYoneda :: Ran Identity f a -> Yoneda f a

as a right Kan lift

yonedaToRift :: Yoneda f a -> Rift Identity f a

Yoneda f can be viewed as the right Kan lift of f along the Identity functor.

 yonedaToRift . riftToYoneda ≡ id
 riftToYoneda . yonedaToRift ≡ id

riftToYoneda :: Rift Identity f a -> Yoneda f a