Data.Functor.Yoneda.Reduction

Description

Yoneda Reduction:

http://ncatlab.org/nlab/show/Yoneda+reduction

Yoneda f is isomorphic to Lan f Identity

Documentation

data Yoneda f a where

A form suitable for Yoneda reduction

Constructors

Yoneda :: (b -> a) -> f b -> Yoneda f a

liftYoneda :: f a -> Yoneda f a

Yoneda expansion

 liftYoneda . lowerYoneda ≡ id
 lowerYoneda . liftYoneda ≡ id

 lift = liftYoneda

lowerYoneda :: Functor f => Yoneda f a -> f a

Yoneda reduction

 lower = lowerM = lowerYoneda

lowerM :: Monad f => Yoneda f a -> f a

Yoneda reduction given a Monad.

 lower = lowerM = lowerYoneda

as a Left Kan extension

yonedaToLan :: Yoneda f a -> Lan Identity f a

Yoneda f is the left Kan extension of f along the Identity functor.

 yonedaToLan . lanToYoneda ≡ id
 lanToYoneda . yonedaToLan ≡ id

lanToYoneda :: Lan Identity f a -> Yoneda f a

as a Left Kan lift

yonedaToLift :: Yoneda f a -> Lift Identity f a

Yoneda f is the left Kan lift of f along the Identity functor.

 yonedaToLift . liftToYoneda ≡ id
 liftToYoneda . yonedaToLift ≡ id

liftToYoneda :: Functor f => Lift Identity f a -> Yoneda f a