Data.Functor.Yoneda

Description

The co-Yoneda lemma states that

f a is isomorphic to (forall r. (a -> r) -> f a)

This is described in a rather intuitive fashion by Dan Piponi in

http://blog.sigfpe.com/2006/11/yoneda-lemma.html

Documentation

newtype Yoneda f a

Constructors

Yoneda

Fields

runYoneda :: forall b. (a -> b) -> f b

liftYoneda :: Functor f => f a -> Yoneda f a

lowerYoneda :: Yoneda f a -> f a

maxF :: (Functor f, Ord (f a)) => Yoneda f a -> Yoneda f a -> Yoneda f a

minF :: (Functor f, Ord (f a)) => Yoneda f a -> Yoneda f a -> Yoneda f a

maxM :: (Monad m, Ord (m a)) => Yoneda m a -> Yoneda m a -> Yoneda m a

minM :: (Monad m, Ord (m a)) => Yoneda m a -> Yoneda m a -> Yoneda m a

as a right Kan extension

yonedaToRan :: Yoneda f a -> Ran Identity f a

Yoneda f is the right Kan extension of f along the Identity functor.

 yonedaToRan . ranToYoneda ≡ id
 ranToYoneda . yonedaToRan ≡ id

ranToYoneda :: Ran Identity f a -> Yoneda f a

as a right Kan lift

yonedaToRift :: Yoneda f a -> Rift Identity f a

Yoneda f is the right Kan lift of f along the Identity functor.

 yonedaToRift . riftToYoneda ≡ id
 riftToYoneda . yonedaToRift ≡ id

riftToYoneda :: Rift Identity f a -> Yoneda f a